refactor to use Game provider
