+++ src/components/App.js
@@ -1,12 +1,22 @@
-import React from "react";
+import React, { useContext } from "react";
 import { BrowserRouter as Router, Route } from "react-router-dom";
 
 import GlobalStyles from "./GlobalStyles";
-import { items } from "../data";
 import Home from "./Home";
 import Game from "./Game";
+import useInterval from "../hooks/use-interval.hook";
+
+import { GameContext } from "./GameContext";
 
 function App(props) {
+  const { numCookies, setNumCookies, cookiesPerSecond } = useContext(
+    GameContext
+  );
+
+  useInterval(() => {
+    setNumCookies(numCookies + cookiesPerSecond);
+  }, 1000);
+
   return (
     <>
       <GlobalStyles />
