+++ src/index.js
@@ -1,8 +1,14 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import React from "react";
+import ReactDOM from "react-dom";
+import { GameProvider } from "./components/GameContext";
 
-import App from './components/App';
+import App from "./components/App";
 
-const rootElement = document.getElementById('root');
+const rootElement = document.getElementById("root");
 
-ReactDOM.render(<App />, rootElement);
+ReactDOM.render(
+  <GameProvider>
+    <App />
+  </GameProvider>,
+  rootElement
+);
